// packages/response-types/src/generated/douyin/Search/Unknown.ts
export type SearchUnknown = {
  __search_type?: never
  ad_info?: { [property: string]: any }
  aweme_list?: null
  backtrace?: string
  challenge_list?: null
  cursor: number
  data?: Data[]
  douyin_ai_search_info?: DouyinAiSearchInfo
  extra: Extra
  global_doodle_config: GlobalDoodleConfig
  guide_search_words?: GuideSearchWord[] | null
  has_more: number
  input_keyword?: string
  log_pb: LogPb
  mock_recall_path?: string
  multi_columns_info?: MultiColumnsInfo
  music_list?: null
  myself_user_id?: string
  ops?: null
  path?: string
  polling_time?: number
  qc?: string
  result_status?: number
  rid?: string
  search_nil_info?: SearchNilInfo
  status_code: number
  time_cost?: TimeCost
  type?: number
  user_list?: UserList3[]
  [property: string]: any
}

type Data = {
  aweme_info?: AwemeInfo
  aweme_list: null
  card_id?: string
  card_info?: CardInfo
  card_style_type?: number
  card_type?: number
  card_type_id?: CardTypeId
  card_unique_name?: string
  data?: Data2
  debug_data?: DebugData
  debug_diff_info?: { [property: string]: any }
  doc_type?: number
  ecom_goods_list: null
  fp?: string
  log_data?: LogData
  lynx_info?: LynxInfo
  music_info_list: null
  ops: null
  provider_doc_id?: number
  provider_doc_id_str?: string
  qishui_music_list: null
  related_word_list?: RelatedWordList[]
  send_back?: string
  shoot_position_list: null
  show_tab: null
  sub_card_list: unknown[] | null
  tab: null
  type: number
  user_list?: UserList2[]
  words_query_record?: WordsQueryRecord
  [property: string]: any
}

type AwemeInfo = {
  ai_follow_images: null
  anchor_info?: AnchorInfo
  anchors: null
  author: Author
  author_user_id: number
  aweme_id: string
  aweme_type: number
  cha_list: null
  challenge_position: null
  chapter_bar_color: null
  chapter_list: null
  collect_stat: number
  comment_list: null
  commerce_config_data: null
  common_left_top_labels: null
  cover_labels: null
  create_scale_type: null
  create_time: number
  danmaku_control?: DanmakuControl
  desc: string
  dislike_dimension_list: null
  dislike_dimension_list_v2: null
  diversion_bar_info: null
  douyin_p_c_video_extra?: string
  effect_inflow_effects: null
  encrypt_interest_point_list: null
  encrypt_key_phrase_list: null
  entertainment_product_info: EntertainmentProductInfo
  fake_horizontal_info?: FakeHorizontalInfo
  follow_shot_assets: null
  geofencing: null
  geofencing_regions: null
  group_id: string
  hybrid_label: null
  image_follow_shot_assets: null
  image_infos: null
  image_list: null
  images: Image[] | null
  img_bitrate: null
  impression_data: ImpressionData
  interaction_stickers: null
  interest_points: null
  is_top: number
  jump_tab_info_list: null
  label_top_text: null
  long_video: null
  media_type: number
  mix_info?: MixInfo
  music: Music
  mv_info: null
  nearby_hot_comment: null
  nickname_position: null
  origin_comment_ids: null
  origin_text_extra: null
  original_images: null
  packed_clips: null
  position: null
  prevent_download: boolean
  promotions: null
  rawdata: string
  ref_tts_id_list: null
  ref_voice_modify_id_list: null
  relation_label?: RelationLabel
  relation_labels: null
  reply_smart_emojis: null
  risk_infos: RiskInfos
  series_info?: SeriesInfo
  share_info: ShareInfo3
  slides_music_beats: null
  social_tag_list: null
  standard_bar_info_list: null
  statistics: Statistics
  status: Status3
  suggest_words: SuggestWords
  text_extra: TextExtra[]
  trends_infos: null
  tts_id_list: null
  uniqid_position: null
  user_digged: number
  video: Video
  video_control: VideoControl
  video_labels: null
  video_tag: null
  video_text: null
  voice_modify_id_list: null
  yumme_recreason: null
  [property: string]: any
}
// ...
type CardInfo = {
  fingerprint: string
  first_pos: number
  index: number
  pos: number
  [property: string]: any
}

type CardTypeId = {
  card_name: string
  doc_type?: number
  performance_infos?: PerformanceInfos
  [property: string]: any
}

type PerformanceInfos = {
  '108': number
  double_column: number
  is_native: number
  search_card_count: number
  [property: string]: any
}

type Data2 = {
  card_tags: null
  common_tab_config: null
  hotspot_tab_config: null
  rs_extra_info: RsExtraInfo
  [property: string]: any
}

type RsExtraInfo = {
  hide_related_words: null
  hide_related_words_id: null
  impr_extra: string
  query_id: string
  [property: string]: any
}

type DebugData = {
  filter_debug_info_list: null
  [property: string]: any
}

type LogData = {
  dcm: string
  search_result_id: string
  [property: string]: any
}

type LynxInfo = {
  is_subcard: boolean
  [property: string]: any
}

type RelatedWordList = {
  extra_info: ExtraInfo
  rand_num: number
  related_img: string
  related_word: string
  word_record: WordRecord
  [property: string]: any
}

type ExtraInfo = {
  words_type: string
  [property: string]: any
}

type WordRecord = {
  force_update_rank: boolean
  group_id: string
  product_id: string
  words_content: string
  words_image: WordsImage
  words_position: number
  words_source: string
  [property: string]: any
}

type WordsImage = {
  url_list: string[]
  [property: string]: any
}

type UserList2 = {
  baikes: null
  challenges: null
  ecom_info: EcomInfo
  effects: null
  fandoms: null
  is_red_uniqueid: boolean
  items: Item[]
  mix_list: null
  musics: null
  position: null
  prefer_profile_tab_type: number
  product_info: null
  product_list: null
  shop_product_info: null
  sub_card_id: SubCardId
  uniqid_position: null
  userSubLightApp: null
  user_info: UserInfo
  user_service_info: unknown[]
  [property: string]: any
}
// ...
type WordsQueryRecord = {
  info: string
  query_id: string
  words_source: string
  [property: string]: any
}

type DouyinAiSearchInfo = {
  ai_search_req_patch: { [property: string]: any }
  is_hit_high_risk: boolean
  is_simple_qa_intent: boolean
  [property: string]: any
}

type Extra = {
  fatal_item_ids: unknown[]
  logid: string
  now: number
  scenes: null
  search_request_id: string
  [property: string]: any
}

type GlobalDoodleConfig = {
  filter_settings: FilterSetting[]
  filter_show_dot?: number
  keyword: string
  [property: string]: any
}

type FilterSetting = {
  android_version?: number
  btm?: string
  default_index: number
  enable_huo_shan?: boolean
  enable_lite?: boolean
  filter_style?: number
  harmony_version?: number
  huoshan_android_version?: number
  huoshan_ios_version?: number
  ios_version?: number
  items: Item2[]
  lite_android_version?: number
  lite_harmony_version?: number
  lite_ios_version?: number
  log_name: string
  name: string
  search_less_text?: SearchLessText
  search_nil_text?: SearchLessText
  title: string
  [property: string]: any
}

type Item2 = {
  log_value: string
  show_dot?: number
  title: string
  value: string
  [property: string]: any
}

type SearchLessText = {
  info: string
  jump_text: string
  [property: string]: any
}

type GuideSearchWord = {
  attached_text: null
  id: string
  query_id: string
  type: string
  word: string
  [property: string]: any
}

type LogPb = {
  impr_id: string
  [property: string]: any
}

type MultiColumnsInfo = {
  group_tag: string
  is_multi_columns: boolean
  [property: string]: any
}

type SearchNilInfo = {
  is_load_more: string
  search_nil_item: string
  search_nil_type: string
  text_type: number
  [property: string]: any
}

type TimeCost = {
  '1st_screen_state': string
  FEDERATION_federation_latency: number
  ack_status: string
  can_ack: string
  complete_time: number
  disable_forecast_cut: number
  disable_optimize_2023: string
  double_column: boolean
  douyin_disease_aggr_video_query_tag: string
  enable_ab_refactor: string
  engine_intent_cat: string
  fed_cache_status: string
  fed_recall_cost: number
  fed_recall_end: number
  first_no_ad: string
  first_screen_card_businesses: string
  first_screen_card_cnt: number
  first_screen_card_names: string
  first_screen_real_lynx_num: number
  first_screen_real_lynx_type: string
  first_screen_sub_card_real_lynx_num: number
  first_screen_sub_card_real_lynx_type: string
  forecast: string
  forecast_ack_reason: string
  forecast_fs_all_hit: string
  forecast_produce_end: number
  forecast_provider_end: number
  goods_card_has_ad: string
  has_ad: string
  hit_user_history_cache: string
  idc: string
  is_cpt_ad: string
  is_go_forecast: string
  is_nil_search: string
  keypoint: string
  loader_time_cost_slowest_loader: string
  loader_time_cost_slowest_loader_latency: number
  log_id: string
  merge_fs_status: string
  merged_fs_all_hit: string
  mf2_append_doc_len: number
  multi_chunk: string
  new_chunk_state: string
  pd: string
  predict_n: number
  preload_miss_count: number
  query: string
  query_tags: string
  quick_ack_failed_detail: string
  reason: string
  search_source_api: string
  server_key_path: string
  skip_provider_ack_reason: string
  stream: string
  stream_inner: number
  target_query_tags: string
  tp_end_forecast_produce: number
  [property: string]: any
}

type UserList3 = {
  baikes: null
  challenges: null
  effects: null
  fandoms: null
  is_red_uniqueid: boolean
  items: null
  mix_list: null
  musics: null
  position: null
  product_info: null
  product_list: null
  shop_product_info: null
  uniqid_position: null
  userSubLightApp: null
  user_info: UserInfo2
  user_service_info: null
  [property: string]: any
}

type UserInfo2 = {
  account_cert_info?: string
  ad_cover_url: null
  avatar_schema_list: null
  avatar_thumb: Avatar168x168
  ban_user_functions: null
  batch_unfollow_contain_tabs: null
  batch_unfollow_relation_desc: null
  can_set_geofencing: null
  card_entries: null
  card_entries_not_display: null
  card_sort_priority: null
  cf_list: null
  cha_list: null
  contrail_list: null
  cover_url: null
  creator_tag_list: null
  custom_verify: string
  data_label_list: null
  display_info: null
  endorsement_info_list: null
  enterprise_verify_reason: string
  familiar_visitor_user: null
  follow_status: number
  follower_count: number
  follower_count_str: string
  follower_list_secondary_information_struct: null
  follower_status: number
  followers_detail: null
  geofencing: null
  homepage_bottom_toast: null
  identity_labels: null
  im_role_ids: null
  interest_tags: null
  item_list: null
  link_item_list: null
  need_points: null
  new_story_cover: null
  nickname: string
  not_seen_item_id_list: null
  not_seen_item_id_list_v2: null
  offline_info_list: null
  personal_tag_list: null
  platform_sync_info: null
  private_relation_list: null
  profile_component_disabled: null
  profile_mob_params: null
  profile_signature_components: null
  relative_users: null
  room_data: string
  room_id?: number
  room_id_str?: string
  sec_uid: string
  secret: number
  short_id: string
  signature: string
  signature_extra: null
  special_people_labels: null
  text_extra: null
  total_favorited: number
  type_label: null
  uid: string
  unique_id: string
  user_permissions: null
  user_tags: UserTag[]
  verification_permission_ids: null
  versatile_display: string
  webcast_preview_labels: null
  white_cover_url: null
  [property: string]: any
}